RealEstate Generation 2.1
RealEstate G2.1 is a minor upgrade of RealEstate G2 that adds volume and mute or print and E-stop button, depending on how you use them to the hatch of the display. A new version of assembly manual adds wiring guides for G2.1 hatch. G2.1 hatch works with G2 display mounts which are unchanged.
This also removes G1 components which is no longer maintained after G2.
This also changes the logo for the whole mod project and move RealEstate STLs models away from models parent folder.