[8-1-bspp-adap-runout] Filament grinding fix
This change fixes one of the problem that lead to permanent damage and false-triggering due to grinding by sliding filament.
By addressing this we added:
* an additional ball bearing,
* an additional M5*30 screw.
The CAD for ELG2 is updated and the BOM has added the materials for such coupler mount with filament runout detector, and is now the recommended config.
Beside this the release status for ELG2 in root readme page has changed from EA to GA, I forgot about this during the official release of it.